test: add internal folder to unit test paths
